+++ xclone-backend/config/urls.py
@@ -18,12 +18,14 @@ from django.contrib import admin
 from django.urls import path, include
 from api import views
 from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView
+from rest_framework.authtoken.views import obtain_auth_token
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     # path('/', views.index, name='index'),
     path('api/', include('api.urls')),
     path('api-auth/', include('rest_framework.urls')),
+    path('api/auth-token/', obtain_auth_token, name='api_token_auth'),
 
      path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
     # Optional UI: